sql-python-trainer/sql-python-trainer/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.26.2
